refactor: unify route structure and create DTOs
- Merge route/ into routes/ folder
- Standardize route file naming to *Router.ts pattern
- Register all routes in router index
- Create DTOs for all entities
- Clean up legacy/test files
